MedEasyInV2

+++ MedEasyIn/src/main/java/com/app/entities/BaseEntity.java
@@ -9,22 +9,22 @@ import javax.persistence.MappedSuperclass;
 public class BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 
 	public BaseEntity() {
 		super();
 	}
 
-	public BaseEntity(int id) {
+	public BaseEntity(Long id) {
 		super();
 		this.id = id;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
